Created Compiler class structure, updated CMakeLists.txt
class Compiler:  created loadFile, which loads the commands vector
with the contents of the file.  Also created printCommands, which
just prints each element of the vector to the console.  Also updated
main.cpp to compile the contents of the file whose name is passed
as a parameter to the command